+++ backend/user_profile/models.py
@@ -112,5 +112,6 @@ def Add_Follow_Notification(sender, instance, action, reverse, pk_set, **kwargs)
         user = User.objects.get(pk=list(pk_set)[0])
         who_following = User.objects.get(username=instance)
         notify = Notification.objects.filter(sender=who_following, user=user, Notification_type=3).first()
-        notify.delete()
+        if notify:
+            notify.delete()
